# docs/openapi.swagger.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "AI Audit System API",
                 "version":  "2.3.0",
                 "description":  "REST API for AI-powered quality audit system with GPT-5",
                 "contact":  {
                                 "name":  "API Support"
                             }
             },
    "servers":  [
                    {
                        "url":  "http://localhost:3002",
                        "description":  "Development server"
                    }
                ],
...
    "paths":  {
                  "/api/audit-configs":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "Audit Configs"
                                                                  ],
                                                         "summary":  "List all audit configurations",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "include_inactive",
                                                                                "in":  "query",
                                                                                "schema":  {
                                                                                               "type":  "boolean"
                                                                                           }
                                                                            },
                                                                            {
                                                                                "name":  "include_steps",
                                                                                "in":  "query",
                                                                                "schema":  {
                                                                                               "type":  "boolean"
                                                                                           }
                                                                            },
                                                                            {
                                                                                "name":  "include_stats",
                                                                                "in":  "query",
                                                                                "schema":  {
                                                                                               "type":  "boolean"
                                                                                           }
                                                                            }
                                                                        ]
                                                     },
                                             "post":  {
                                                          "tags":  [
                                                                       "Audit Configs"
                                                                   ],
                                                          "summary":  "Create new audit configuration (atomic transaction)",
                                                          "description":  "Creates audit config with all steps in a single transaction. All or nothing."
                                                      }
                                         },
                  "/api/audit-configs/{id}":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "Audit Configs"
                                                                       ],
                                                              "summary":  "Get audit configuration details",
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "string"
                                                                                                }
                                                                                 },
                                                                                 {
                                                                                     "name":  "include_stats",
                                                                                     "in":  "query",
                                                                                     "schema":  {
                                                                                                    "type":  "boolean"
                                                                                                }
                                                                                 }
                                                                             ]
                                                          },
                                                  "put":  {
                                                              "tags":  [
                                                                           "Audit Configs"
                                                                       ],
                                                              "summary":  "Update audit configuration"
                                                          },
                                                  "delete":  {
                                                                 "tags":  [
                                                                              "Audit Configs"
                                                                          ],
                                                                 "summary":  "Delete audit configuration"
                                                             }
                                              },
...
                  "/api/audits":  {
                                      "get":  {
                                                  "tags":  [
                                                               "Audits"
                                                           ],
                                                  "summary":  "List all audits with advanced filtering and sorting",
                                                  "description":  "Retrieve audits across all fiches with comprehensive filtering options.\nSupports pagination, date ranges, status filtering, and multiple sort options.\n",
                                                  "parameters":  [
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "fiche_ids",
                                                                         "schema":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                         "description":  "Comma-separated list of fiche IDs to filter by",
                                                                         "example":  "1762209,1753254"
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "status",
                                                                         "schema":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                         "description":  "Comma-separated list of statuses (pending, running, completed, failed)",
                                                                         "example":  "completed,failed"
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "is_compliant",
                                                                         "schema":  {
                                                                                        "type":  "boolean"
                                                                                    },
                                                                         "description":  "Filter by compliance status",
                                                                         "example":  true
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "date_from",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "format":  "date"
                                                                                    },
                                                                         "description":  "Start date (ISO 8601)",
                                                                         "example":  "2025-01-01"
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "date_to",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "format":  "date"
                                                                                    },
                                                                         "description":  "End date (ISO 8601)",
                                                                         "example":  "2025-01-31"
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "audit_config_ids",
                                                                         "schema":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                         "description":  "Comma-separated list of audit config IDs",
                                                                         "example":  "13,11"
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "sort_by",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "enum":  [
                                                                                                     "created_at",
                                                                                                     "completed_at",
                                                                                                     "score_percentage",
                                                                                                     "duration_ms"
                                                                                                 ]
                                                                                    },
                                                                         "description":  "Field to sort by",
                                                                         "example":  "created_at"
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "sort_order",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "enum":  [
                                                                                                     "asc",
                                                                                                     "desc"
                                                                                                 ]
                                                                                    },
                                                                         "description":  "Sort order",
                                                                         "example":  "desc"
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "limit",
                                                                         "schema":  {
                                                                                        "type":  "integer"
                                                                                    },
                                                                         "description":  "Number of results per page (max 500)",
                                                                         "example":  50
                                                                     },
                                                                     {
                                                                         "in":  "query",
                                                                         "name":  "offset",
                                                                         "schema":  {
                                                                                        "type":  "integer"
                                                                                    },
                                                                         "description":  "Number of results to skip for pagination",
                                                                         "example":  0
                                                                     }
                                                                 ],
                                                  "responses":  {
                                                                    "200":  {
                                                                                "description":  "List of audits with pagination info"
                                                                            },
                                                                    "400":  {
                                                                                "description":  "Invalid query parameters"
                                                                            },
                                                                    "500":  {
                                                                                "description":  "Server error"
                                                                            }
                                                                }
                                              }
                                  },
...
              },
...
}